requirements.txt
sounddevice
soundfile
numpy
vosk
pyusb==1.2.1